+++ object_parsing/object_parsing.cpp
@@ -1,4 +1,3 @@
-#include "stdafx.h"
 #include "object_parsing.hpp"
 #include "Parsers\parsers.hpp"
 
@@ -6,5 +5,5 @@ std::string associated_string(const token_group& left, const token_group* child_
 	if((left.association != association_type::eq) | (left.group_size == 0) | (child_start + 1 != child_end))
 		return std::string("");
 
-	return std::string(child_start->token.start, child_start->token.end - child_start->token.start);
+	return std::string(child_start->token.start, static_cast<size_t>(child_start->token.end - child_start->token.start));
 }
